feat: opentelemetry integration, removed self designed tracing
