+++ core/trace/config.go
@@ -1,4 +1,4 @@
-package opentelemetry
+package trace
 
 // TraceName represents the tracing name.
 const TraceName = "go-zero"
@@ -8,5 +8,5 @@ type Config struct {
 	Name     string  `json:",optional"`
 	Endpoint string  `json:",optional"`
 	Sampler  float64 `json:",default=1.0"`
-	Batcher  string  `json:",default=jaeger"`
+	Batcher  string  `json:",default=jaeger,options=jaeger"`
 }
