--- core/trace/constants.go
@@ -1,8 +1,0 @@
-package trace
-
-const (
-	// TraceIdKey is the trace id header.
-	TraceIdKey = "X-Trace-ID"
-
-	spanIdKey = "X-Span-ID"
-)
