+++ core/trace/message.go
@@ -1,4 +1,4 @@
-package opentelemetry
+package trace
 
 import (
 	"context"
