+++ core/trace/vars.go
@@ -1,0 +1,8 @@
+package trace
+
+import "net/http"
+
+// TraceIdKey is the trace id header.
+// https://www.w3.org/TR/trace-context/#trace-id
+// May change it to trace-id afterwards.
+var TraceIdKey = http.CanonicalHeaderKey("x-Trace-id")
