+++ zrpc/internal/client.go
@@ -67,16 +67,14 @@ func (c *client) buildDialOptions(opts ...ClientOption) []grpc.DialOption {
 		grpc.WithInsecure(),
 		grpc.WithBlock(),
 		WithUnaryClientInterceptors(
-			clientinterceptors.UnaryTracingInterceptor,
-			clientinterceptors.UnaryOpenTracingInterceptor(),
+			clientinterceptors.UnaryTracingInterceptor(),
 			clientinterceptors.DurationInterceptor,
 			clientinterceptors.PrometheusInterceptor,
 			clientinterceptors.BreakerInterceptor,
 			clientinterceptors.TimeoutInterceptor(cliOpts.Timeout),
 		),
 		WithStreamClientInterceptors(
-			clientinterceptors.StreamTracingInterceptor,
-			clientinterceptors.StreamOpenTracingInterceptor(),
+			clientinterceptors.StreamTracingInterceptor(),
 		),
 	}
