+++ zrpc/internal/clientinterceptors/tracinginterceptor_test.go
@@ -9,15 +9,31 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/tal-tech/go-zero/core/trace"
 	"google.golang.org/grpc"
-	"google.golang.org/grpc/metadata"
 )
+
+func TestOpenTracingInterceptor(t *testing.T) {
+	trace.StartAgent(trace.Config{
+		Name:     "go-zero-test",
+		Endpoint: "http://localhost:14268/api/traces",
+		Batcher:  "jaeger",
+		Sampler:  1.0,
+	})
+
+	cc := new(grpc.ClientConn)
+	err := UnaryTracingInterceptor()(context.Background(), "/ListUser", nil, nil, cc,
+		func(ctx context.Context, method string, req, reply interface{}, cc *grpc.ClientConn,
+			opts ...grpc.CallOption) error {
+			return nil
+		})
+	assert.Nil(t, err)
+}
 
 func TestUnaryTracingInterceptor(t *testing.T) {
 	var run int32
 	var wg sync.WaitGroup
 	wg.Add(1)
 	cc := new(grpc.ClientConn)
-	err := UnaryTracingInterceptor(context.Background(), "/foo", nil, nil, cc,
+	err := UnaryTracingInterceptor()(context.Background(), "/foo", nil, nil, cc,
 		func(ctx context.Context, method string, req, reply interface{}, cc *grpc.ClientConn,
 			opts ...grpc.CallOption) error {
 			defer wg.Done()
@@ -34,7 +50,7 @@ func TestStreamTracingInterceptor(t *testing.T) {
 	var wg sync.WaitGroup
 	wg.Add(1)
 	cc := new(grpc.ClientConn)
-	_, err := StreamTracingInterceptor(context.Background(), nil, cc, "/foo",
+	_, err := StreamTracingInterceptor()(context.Background(), nil, cc, "/foo",
 		func(ctx context.Context, desc *grpc.StreamDesc, cc *grpc.ClientConn, method string,
 			opts ...grpc.CallOption) (grpc.ClientStream, error) {
 			defer wg.Done()
@@ -50,14 +66,8 @@ func TestUnaryTracingInterceptor_GrpcFormat(t *testing.T) {
 	var run int32
 	var wg sync.WaitGroup
 	wg.Add(1)
-	md := metadata.New(map[string]string{
-		"foo": "bar",
-	})
-	carrier, err := trace.Inject(trace.GrpcFormat, md)
-	assert.Nil(t, err)
-	ctx, _ := trace.StartServerSpan(context.Background(), carrier, "user", "/foo")
 	cc := new(grpc.ClientConn)
-	err = UnaryTracingInterceptor(ctx, "/foo", nil, nil, cc,
+	err := UnaryTracingInterceptor()(context.Background(), "/foo", nil, nil, cc,
 		func(ctx context.Context, method string, req, reply interface{}, cc *grpc.ClientConn,
 			opts ...grpc.CallOption) error {
 			defer wg.Done()
@@ -73,14 +83,8 @@ func TestStreamTracingInterceptor_GrpcFormat(t *testing.T) {
 	var run int32
 	var wg sync.WaitGroup
 	wg.Add(1)
-	md := metadata.New(map[string]string{
-		"foo": "bar",
-	})
-	carrier, err := trace.Inject(trace.GrpcFormat, md)
-	assert.Nil(t, err)
-	ctx, _ := trace.StartServerSpan(context.Background(), carrier, "user", "/foo")
 	cc := new(grpc.ClientConn)
-	_, err = StreamTracingInterceptor(ctx, nil, cc, "/foo",
+	_, err := StreamTracingInterceptor()(context.Background(), nil, cc, "/foo",
 		func(ctx context.Context, desc *grpc.StreamDesc, cc *grpc.ClientConn, method string,
 			opts ...grpc.CallOption) (grpc.ClientStream, error) {
 			defer wg.Done()
