+++ zrpc/internal/serverinterceptors/tracinginterceptor_test.go
@@ -7,13 +7,39 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/tal-tech/go-zero/core/trace/tracespec"
+	"github.com/tal-tech/go-zero/core/trace"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/metadata"
 )
 
+func TestUnaryOpenTracingInterceptor_Disable(t *testing.T) {
+	interceptor := UnaryTracingInterceptor()
+	_, err := interceptor(context.Background(), nil, &grpc.UnaryServerInfo{
+		FullMethod: "/",
+	}, func(ctx context.Context, req interface{}) (interface{}, error) {
+		return nil, nil
+	})
+	assert.Nil(t, err)
+}
+
+func TestUnaryOpenTracingInterceptor_Enabled(t *testing.T) {
+	trace.StartAgent(trace.Config{
+		Name:     "go-zero-test",
+		Endpoint: "http://localhost:14268/api/traces",
+		Batcher:  "jaeger",
+		Sampler:  1.0,
+	})
+	interceptor := UnaryTracingInterceptor()
+	_, err := interceptor(context.Background(), nil, &grpc.UnaryServerInfo{
+		FullMethod: "/package.TestService.GetUser",
+	}, func(ctx context.Context, req interface{}) (interface{}, error) {
+		return nil, nil
+	})
+	assert.Nil(t, err)
+}
+
 func TestUnaryTracingInterceptor(t *testing.T) {
-	interceptor := UnaryTracingInterceptor("foo")
+	interceptor := UnaryTracingInterceptor()
 	var run int32
 	var wg sync.WaitGroup
 	wg.Add(1)
@@ -29,42 +55,8 @@ func TestUnaryTracingInterceptor(t *testing.T) {
 	assert.Equal(t, int32(1), atomic.LoadInt32(&run))
 }
 
-func TestUnaryTracingInterceptor_GrpcFormat(t *testing.T) {
-	interceptor := UnaryTracingInterceptor("foo")
-	var wg sync.WaitGroup
-	wg.Add(1)
-	var md metadata.MD
-	ctx := metadata.NewIncomingContext(context.Background(), md)
-	_, err := interceptor(ctx, nil, &grpc.UnaryServerInfo{
-		FullMethod: "/",
-	}, func(ctx context.Context, req interface{}) (interface{}, error) {
-		defer wg.Done()
-		assert.True(t, len(ctx.Value(tracespec.TracingKey).(tracespec.Trace).TraceId()) > 0)
-		assert.True(t, len(ctx.Value(tracespec.TracingKey).(tracespec.Trace).SpanId()) > 0)
-		return nil, nil
-	})
-	wg.Wait()
-	assert.Nil(t, err)
-}
-
-func TestStreamTracingInterceptor(t *testing.T) {
-	interceptor := StreamTracingInterceptor("foo")
-	var run int32
-	var wg sync.WaitGroup
-	wg.Add(1)
-	err := interceptor(nil, new(mockedServerStream), nil,
-		func(srv interface{}, stream grpc.ServerStream) error {
-			defer wg.Done()
-			atomic.AddInt32(&run, 1)
-			return nil
-		})
-	wg.Wait()
-	assert.Nil(t, err)
-	assert.Equal(t, int32(1), atomic.LoadInt32(&run))
-}
-
 func TestStreamTracingInterceptor_GrpcFormat(t *testing.T) {
-	interceptor := StreamTracingInterceptor("foo")
+	interceptor := StreamTracingInterceptor()
 	var run int32
 	var wg sync.WaitGroup
 	wg.Add(1)
